Add date and date range picker components with a string state
